updates entrypoint slightly

+++ entrypoint.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 set -e
 
-if [[ ! -z "$AWS_REGION" && -z "$DISABLE_AWS_SECRETS" ]]; then
-  # specific code for AWS secrets manager, if being used. runs get_secrets.rb
-  # to create a temp env var file, then sets envs, then deletes file
+if [[ ! -z "$AWS_REGION" && ! -z "$AWS_SECRETS_PREFIX" && -z "$DISABLE_AWS_SECRETS" ]]; then
+  # Specific code for AWS secrets manager, if being used. Runs get_secrets.rb
+  # to create a temp env var file, then sets envs, then deletes file.
   /usr/bin/set_secrets.rb
   eval $(cat /tmp/secrets.env | sed 's/^/export /')
   rm -f /tmp/secrets.env
@@ -16,14 +16,17 @@ fi
 
 echo "Checking if database exists…"
 
-DB_NAME=$DB_NAME
+CURR_DB_NAME=$DB_NAME
 
 if [[ "$RAILS_ENV" != "production" && "$RAILS_ENV" != "prod" ]]; then
-  DB_NAME="${DB_NAME}-${RAILS_ENV}"
+  CURR_DB_NAME="${DB_NAME}-${RAILS_ENV}"
 fi
 
-# then pass vars to psql, which will test whether the DB exists
-if echo "\c $DB_NAME; \dt" | psql -h "$DB_HOST" -U "$DB_USER" -d "$DB_NAME" | grep schema_migrations 2>&1 >/dev/null
+# Need to declare a PGPASSWORD variable for psql, which expects it.
+PGPASSWORD=${DB_PASSWORD}
+
+# Then pass vars to psql, which will test whether the DB exists
+if echo "\c $CURR_DB_NAME; \dt" | psql -h "$DB_HOST" -U "$DB_USERNAME" -d "$CURR_DB_NAME" | grep schema_migrations 2>&1 >/dev/null
 then
   echo "Past migrations found. Running any new migrations…"
   bundle exec rake db:migrate
@@ -36,5 +39,5 @@ else
   fi
 fi
 
-# Then exec the container's main process (what's set as CMD in the Dockerfile).
+# Finally, exec the container's main process (what's set as CMD in the Dockerfile).
 exec "$@"
